// movil/src/app/models/trazabilidad/trazabilidad.response.ts
// Generated by https://quicktype.io

export interface TrazabilidadResponse {
    code:         number;
    status:       string;
    distribucion: Distribucion[];
}

export interface Distribucion {
    contacto:     Contacto;
    trazabilidad: Trazabilidad[];
    InfoDespacho: InfoDespacho[];
}

export interface InfoDespacho {
    Cantidad:       number;
    caja_numero:    number;
    bandeja_numero: number;
}

export interface Contacto {
    Remision:      string;
    FechaEntrega:  string;
    Cliente:       string;
    Destino:       string;
    Finca:         string;
    Facturado:     number;
    Adicionales:   number;
    Reposicion:    number;
    Total:         number;
    Total_enviado: number;
    Maximo:        number;
}

export interface Trazabilidad {
    NumLote:       string;
    Fechadesove:   string;
    LineaGenetica: string;
    edad:          number;
    tamanio:       number;
    ovas_ml:       number;
}


/*** clase adicional para organizar la informacion */
export class InfoDespachoMostrar {
    constructor(cantidad, cajanumero, bandeja){
        this.Cantidad = cantidad;
        this.caja_numero = cajanumero;
        this.bandeja_numero = bandeja;

    }
    Cantidad:       string;
    caja_numero:    string;
    bandeja_numero: string;
}